requirements.txt
flask==3.0.3
python-dateutil==2.9.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.1
flask-migrate==4.0.7
apscheduler==3.10.4
python-dotenv==1.0.1
werkzeug==3.0.3
requests==2.32.3
pandas==2.2.2
openpyxl==3.1.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
gunicorn==22.0.0
intuitlib==1.3.0
google-ads==24.1.0
cryptography==42.0.8
